Move test colletor implementation from api
