+++ engine/flink/process/src/main/scala/pl/touk/nussknacker/engine/process/registrar/CollectingSinkFunction.scala
@@ -2,11 +2,9 @@ package pl.touk.nussknacker.engine.process.registrar
 
 import org.apache.flink.streaming.api.functions.sink.{RichSinkFunction, SinkFunction}
 import pl.touk.nussknacker.engine.api.InterpretationResult
-import pl.touk.nussknacker.engine.api.test.InvocationCollectors.SinkInvocationCollector
-import pl.touk.nussknacker.engine.compiledgraph.part.SinkPart
-import pl.touk.nussknacker.engine.graph.node
-import pl.touk.nussknacker.engine.process.{ExceptionHandlerFunction, ProcessPartFunction}
+import pl.touk.nussknacker.engine.process.ExceptionHandlerFunction
 import pl.touk.nussknacker.engine.process.compiler.FlinkProcessCompilerData
+import pl.touk.nussknacker.engine.testmode.Collectors.SinkInvocationCollector
 
 private[registrar] class CollectingSinkFunction(val compiledProcessWithDepsProvider: ClassLoader => FlinkProcessCompilerData,
                                                 collectingSink: SinkInvocationCollector, sinkId: String)
