+++ engine/interpreter/src/main/scala/pl/touk/nussknacker/engine/testmode/Collectors.scala
@@ -1,0 +1,109 @@
+package pl.touk.nussknacker.engine.testmode
+
+import pl.touk.nussknacker.engine.api.{ContextId, InterpretationResult}
+import pl.touk.nussknacker.engine.api.context.ProcessCompilationError.NodeId
+import pl.touk.nussknacker.engine.api.test.InvocationCollectors.{CollectableAction, ServiceInvocationCollector, ToCollect, TransmissionNames}
+import pl.touk.nussknacker.engine.testmode.Collectors.CollectorMode.CollectorMode
+
+import scala.concurrent.{ExecutionContext, Future}
+
+object Collectors {
+
+  type ServiceInvocationCollectorForContext = ContextId => ServiceInvocationCollector
+
+  case class QueryServiceResult(name: String, result: Any)
+
+  private[testmode] object CollectorMode extends Enumeration {
+    type CollectorMode = Value
+    val Test, Query, Production = Value
+  }
+
+  trait BaseServiceInvocationCollector extends ServiceInvocationCollector {
+
+    protected def runIdOpt: Option[TestRunId]
+    protected def collectorMode: CollectorMode
+    protected def updateResult(runId: TestRunId, testInvocation: Any, name: String): Unit
+
+    def collect[A](request: => ToCollect, mockValue: Option[A])(action: => Future[A], names: TransmissionNames = TransmissionNames.default)
+                  (implicit ec: ExecutionContext): Future[A] = {
+      collectWithResponse[A](request, mockValue)(action.map(a => CollectableAction(() => "", a)), names)
+    }
+
+    def collectWithResponse[A](request: => ToCollect, mockValue: Option[A])(action: => Future[CollectableAction[A]], names: TransmissionNames = TransmissionNames.default)
+                              (implicit ec: ExecutionContext): Future[A] = {
+      def enabledMode = if (runIdOpt.isDefined) collectorMode else CollectorMode.Production
+      def runId = runIdOpt.getOrElse(throw new IllegalStateException(s"Collector has to be enabled in $enabledMode"))
+
+      enabledMode match {
+        case CollectorMode.Production =>
+          action.map(_.result)
+        case CollectorMode.Test => mockValue match {
+          case Some(mockVal) =>
+            updateResult(runId, request, names.invocationName)
+            Future.successful(mockVal)
+          case None =>
+            action.map(_.result)
+        }
+        case CollectorMode.Query =>
+          updateResult(runId, request, names.invocationName)
+          action.map { collectableAction =>
+            updateResult(runId, collectableAction.toCollect(), names.resultName)
+            collectableAction.result
+          }
+      }
+    }
+  }
+
+  case class QueryServiceInvocationCollector(runIdOpt: Option[TestRunId], serviceName: String) extends BaseServiceInvocationCollector {
+
+    override protected def collectorMode: CollectorMode = CollectorMode.Query
+
+    override protected def updateResult(runId: TestRunId, testInvocation: Any, name: String): Unit = {
+      val queryServiceResult = QueryServiceResult(name, testInvocation)
+      QueryResultsHolder.updateResult(runId, queryServiceResult)
+    }
+
+    def getResults[T]: List[QueryServiceResult] = {
+      withRunIdDefined { runId =>
+        QueryResultsHolder.queryResultsForId(runId)
+      }
+    }
+
+    def cleanResults(): Unit = {
+      withRunIdDefined { runId =>
+        QueryResultsHolder.cleanResult(runId)
+      }
+    }
+
+    private def withRunIdDefined[T](f: TestRunId => T): T = {
+      runIdOpt match {
+        case Some(runId) => f(runId)
+        case None => throw new IllegalStateException("RunId is not defined")
+      }
+    }
+
+  }
+
+  case class TestServiceInvocationCollector(runIdOpt: Option[TestRunId],
+                                            contextId: ContextId,
+                                            nodeId: NodeId,
+                                            serviceRef: String) extends BaseServiceInvocationCollector {
+
+    override protected def collectorMode: CollectorMode = CollectorMode.Test
+
+    override protected def updateResult(runId: TestRunId, testInvocation: Any, name: String): Unit = {
+      ResultsCollectingListenerHolder.updateResults(
+        runId, _.updateMockedResult(nodeId.id, contextId, serviceRef, testInvocation)
+      )
+    }
+  }
+
+  case class SinkInvocationCollector(runId: TestRunId, nodeId: String, ref: String, outputPreparer: Any => String) {
+
+    def collect(result: InterpretationResult): Unit = {
+      val mockedResult = outputPreparer(result.output)
+      ResultsCollectingListenerHolder.updateResults(runId, _.updateMockedResult(nodeId, ContextId(result.finalContext.id), ref, mockedResult))
+    }
+  }
+
+}
